PCHR-4473: Add views_slideshow as a dependency to be download with drush make
This is a dependency of the onboarding_slideshow feature, in the
civihr_employee_portal module. Since it was not included in the make file,
it would be downloaded during the enabling/installation of the module.

In most of the cases this would work fine, but in some rare and uncertain
circumstances, the installation would fail before finishing the download.
By adding the dependency to the make file, we force it to be downloaded
by the drush make command instead, so that when the installation gets to
the point of enabling the module all the code will already be availble
and it won't fail trying to download it.